backend/requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.0.0
supabase>=1.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
